requirements.txt
kafka-python==2.3.1
snowflake-connector-python==4.4.0
